Sonyc_Backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
langchain-mistralai
langchain-community
langchain-core
langchain-text-splitters
youtube-transcript-api
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
email-validator
chromadb
pypdf
beautifulsoup4
requests
pydantic-settings>=2.0.0
PyGithub







